+++ src/test/scala/satify/DPLLTest/PartialModelUtilsTest.scala
@@ -1,4 +1,4 @@
-package satify.DecisionTreeTest
+package satify.DPLLTest
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
